Refactor click-actions build
Preparing the structure so that including functions to click-actions is
easier/makes sense.

Also it's much better not to have the build files on the same level as
user scripts.
