+++ Makefile
@@ -1,7 +1,9 @@
 CFLAGS = -g -O2 -Wall
 LDFLAGS = -lX11
 
-all: main windowlist.o click-actions/raise click-actions/minimize click-actions/close
+.PHONY: all clean
+
+all: main click-actions/raise click-actions/minimize click-actions/close
 
 main: main.c windowlist.o windowlist.h toml-c.h
 	gcc $(CFLAGS) $(LDFLAGS) -o main main.c windowlist.o
@@ -9,18 +11,17 @@ main: main.c windowlist.o windowlist.h toml-c.h
 windowlist.o: windowlist.c
 	gcc $(CFLAGS) -c windowlist.c
 
-click-actions/raise: click-actions/raise.c
-	gcc $(CFLAGS) $(LDFLAGS) -o click-actions/raise click-actions/raise.c
+click-actions/src/common.o: click-actions/src/common.c
+	gcc $(CFLAGS) -c -o click-actions/src/common.o click-actions/src/common.c
 
-click-actions/minimize: click-actions/minimize.c
-	gcc $(CFLAGS) $(LDFLAGS) -o click-actions/minimize click-actions/minimize.c
-
-click-actions/close: click-actions/close.c
-	gcc $(CFLAGS) $(LDFLAGS) -o click-actions/close click-actions/close.c
+# raise, minimize, close executables
+click-actions/%: click-actions/src/%.c click-actions/src/common.o
+	gcc $(CFLAGS) $(LDFLAGS) -o $@ $^
 
 clean:
-	rm windowlist.o
-	rm main
-	rm click-actions/raise
-	rm click-actions/minimize
-	rm click-actions/close
+	rm -f main \
+	      windowlist.o \
+	      click-actions/src/common.o \
+	      click-actions/raise \
+	      click-actions/minimize \
+	      click-actions/close
