+++ click-actions/src/common.h
@@ -1,0 +1,4 @@
+#include <X11/Xlib.h>
+
+int client_msg(Display* d, Window w, char* msg);
+Window str_to_wid(char* str);
